requirements.txt
# -i http://mirrors.aliyun.com/pypi/simple/  --trusted-host mirrors.aliyun.com
djangorestframework==3.14.0